Refactor tests into the appropriate test classes
